rename: featuers/ktWiz index 관리(#206)
