+++ src/features/ticket/index.ts
@@ -1,1 +1,7 @@
+/** tab */
+export { TicketPriceTab } from './TicketPriceTab';
+export { TicketReservationTab } from './TicketReservationTab';
+
+/** components */
 export { LoginConfirmationDialog } from './components/LoginConfirmationDialog';
+export { TicketItem } from './components/TicketItem';
