+++ src/pages/ticketpurchase/TicketPurchasePage.tsx
@@ -1,7 +1,6 @@
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui';
 import { Banner } from '@/features/common';
-import TicketPriceTab from '@/features/ticket/components/TicketPriceTab';
-import TicketReservationTab from '@/features/ticket/components/TicketReservationTab';
+import { TicketPriceTab, TicketReservationTab } from '@/features/ticket';
 import { useTabFromUrl } from '@/hooks/useTabFromUrl';
 
 const WIZ_TABS_CONFIG = [
